moved functions around
