+++ src/updateLoops.lua
@@ -1,0 +1,82 @@
+function map_loop()
+  if btnp(❎) then 
+    load_game(map_selector.pos + 1)
+    map_menu_enable = false 
+    return
+  end
+
+  local dx, dy = controls()
+  if dx < 0 then 
+    map_selector.pos = (map_selector.pos - 1) % #map_data
+    map_selector.x = shop_ui_data.x[map_selector.pos + 1]-20
+  elseif dx > 0 then 
+    map_selector.pos = (map_selector.pos + 1) % #map_data
+    map_selector.x = shop_ui_data.x[map_selector.pos + 1]-20
+  end
+end
+
+function shop_loop()
+  foreach(menus, Menu.update)
+  
+  if btnp(🅾️) then -- disable shop
+    if get_active_menu().prev == nil then 
+      shop_enable = false
+      menus[1].enable = false
+      return
+    else
+      swap_menu_context(get_active_menu().prev)
+    end
+  end
+  if btnp(❎) then 
+    Menu.invoke(get_active_menu())
+  end
+
+  foreach(menus, Menu.move)
+end
+
+function game_loop()
+  if (auto_start_wave) start_round()
+
+  if btnp(🅾️) then
+    shop_enable = true
+    menus[1].enable = true
+    return
+  end
+  if btnp(❎) then 
+    local position = selector.position/8
+    if is_in_table(position, towers, true) then 
+      refund_tower_at(position)
+    else
+      place_tower(position)
+    end
+  end
+
+  selector.position += Vec:new(controls()) * 8
+  Vec.clamp(selector.position, 0, 120)
+
+  -- update objs
+  if enemies_active then 
+    foreach(enemies, update_enemy_position)
+    foreach(towers, Tower.attack)
+    if start_next_wave then 
+      start_next_wave = false
+      wave_cor = cocreate(spawn_enemy)
+    end
+    if wave_cor and costatus(wave_cor) ~= 'dead' then
+      coresume(wave_cor)
+    else
+      wave_cor = nil
+    end
+  end
+  -- update particles
+  foreach(particles, Particle.tick)
+  foreach(animators, Animator.update)
+  if (not enemies_active and incoming_hint) foreach(incoming_hint, Animator.update)
+  -- clean up tables
+  foreach(enemies, kill_enemy)
+  foreach(particles, destroy_particle)
+  if enemies_active and #enemies == 0 and enemies_remaining == 0 then 
+    enemies_active = false 
+    sfx(sfx_data.round_complete)
+  end
+end
